requirements.txt
python-docx
python-pptx
openpyxl
pdf2image
pytesseract
pdfminer.six